refactor(core): consolidate duplicate functions and split request.ts
- Consolidate isStreamingCompletionModel: export from completion/create-completion.ts, re-export from agent/utils.ts
- Consolidate extractRagText: remove duplicate from extractor/extractor.ts, import from agent/utils.ts
- Consolidate isToolResultContentArray + serializeToolOutput: export from completion/types.ts, import from tool/tool.ts
- Consolidate isRecord/isPlainRecord: create shared utility in internal/compact.ts, update skills/local.ts and mcp/result.ts
- Split request.ts: extract types and helper functions to request-types.ts (878 -> 786 lines)
- All 205 core tests pass, 71 studio tests pass
