+++ packages/core/src/agent/index.ts
@@ -55,4 +55,4 @@ export type {
   AgentChildStreamEvent,
   AgentStreamEvent,
   PromptResponse,
-} from "./request";
+} from "./request-types";
