+++ packages/core/src/agent/request-types.ts
@@ -1,0 +1,106 @@
+import {
+  type CompletionResponse,
+  type Message as MessageType,
+  type ReasoningContentType,
+  type ToolCall,
+  type ToolResultContent,
+  Usage,
+} from "../completion/index";
+import type { AgentTraceInfo } from "../observability/types";
+import type { AgentDeltaEvent } from "./stream-accumulator";
+
+export type PromptResponse = {
+  output: string;
+  usage: Usage;
+  messages: MessageType[];
+  trace?: AgentTraceInfo | undefined;
+};
+
+export type AgentChildStreamEvent<RawResponse = unknown> =
+  | {
+      type: "turn_start";
+      turn: number;
+      prompt: MessageType;
+      history: MessageType[];
+    }
+  | {
+      type: "text_delta";
+      turn: number;
+      delta: string;
+    }
+  | {
+      type: "reasoning_delta";
+      turn: number;
+      delta: string;
+      id?: string;
+      contentType?: ReasoningContentType;
+      signature?: string;
+    }
+  | {
+      type: "tool_call";
+      turn: number;
+      toolCall: ToolCall;
+    }
+  | {
+      type: "tool_result";
+      turn: number;
+      toolName: string;
+      toolCallId?: string;
+      internalCallId: string;
+      args: string;
+      result: string;
+      structuredResult?: ToolResultContent[] | undefined;
+    }
+  | {
+      type: "turn_end";
+      turn: number;
+      response: CompletionResponse<RawResponse>;
+    }
+  | {
+      type: "final";
+      runId: string;
+      output: string;
+      usage: Usage;
+      messages: MessageType[];
+      trace?: AgentTraceInfo | undefined;
+    }
+  | {
+      type: "error";
+      error: unknown;
+    };
+
+export type AgentStreamEvent<RawResponse = unknown> =
+  | AgentChildStreamEvent<RawResponse>
+  | {
+      type: "agent_tool_event";
+      turn: number;
+      toolName: string;
+      toolCallId?: string;
+      internalCallId: string;
+      agentId: string;
+      agentName?: string;
+      event: AgentChildStreamEvent<RawResponse>;
+    };
+
+export function addTurn(turn: number, event: AgentDeltaEvent): AgentStreamEvent {
+  if (event.type === "text_delta") {
+    return { type: "text_delta", turn, delta: event.delta };
+  }
+  if (event.type === "reasoning_delta") {
+    const mapped: AgentStreamEvent = { type: "reasoning_delta", turn, delta: event.delta };
+    if (event.id !== undefined) mapped.id = event.id;
+    if (event.contentType !== undefined) mapped.contentType = event.contentType;
+    if (event.signature !== undefined) mapped.signature = event.signature;
+    return mapped;
+  }
+  return { type: "tool_call", turn, toolCall: event.toolCall };
+}
+
+export function isGenerationDeltaEvent(type: string): boolean {
+  return (
+    type === "text_delta" ||
+    type === "reasoning_delta" ||
+    type === "tool_call_delta" ||
+    type === "tool_call"
+  );
+}
